자잘한 오류 수정

+++ src/main/java/project/app/almond/service/CommLikesService.java
@@ -53,6 +53,6 @@ public class CommLikesService {
 		else if(type==-2) content="["+uvo1.getNickname()+"]¥‘¿Ã 'Ω»æÓø‰'∏¶ √Îº“«œºÃΩ¿¥œ¥Ÿ.";
 		else content="["+uvo1.getNickname()+"]¥‘¿Ã ¥Ò±€¿ª ¥ŸºÃΩ¿¥œ¥Ÿ.";
 		String showdate=new SimpleDateFormat("yyyyMMddHHmmss").format(new java.util.Date().getTime());
-		adao.insert(new AlarmVo(0, uvo2.getUsernum(), title, content, null, showdate,vo.getType(), (Integer)map.get("CONTNUM"), 0));
+		adao.insert(new AlarmVo(0, uvo2.getUsernum(), title, content, null, showdate,vo.getType(), ((BigDecimal)map.get("CONTNUM")).intValue(), 0));
 	}
 }
